Refactor : ID값을 Integer, User 클래스를 Users로 변경

+++ src/main/java/com/kdt/firststep/counselor/controller/PersonalityController.java
@@ -16,7 +16,7 @@ public class PersonalityController {
     private final PersonalityService personalityService;
 
     @GetMapping("/check")
-    public ResponseEntity<PersonalityCheckResponseDto> checkPersonality(@RequestParam Long userId) {
+    public ResponseEntity<PersonalityCheckResponseDto> checkPersonality(@RequestParam Integer userId) {
         return ResponseEntity.ok(personalityService.checkPersonalityStatus(userId));
     }
 }
